21/9
Managed to make the "computer" block everyone of the player attemps,
except for only one. Still is posible to win, but very difficult.

Next updates would consist in tracking the "computer" own moves, so
that if is possible to mark 3 in line instead of blocking the rival,
it would be done.

Consider to reduce difficulty of the game, so that the chance of beating
the "computer" increases notoriously compared to the one now.

Also it would be necesary to make the "computer" choose a position left
to finish a game, even if it means that the game will finish in a draw.

The game should be able to draw a vertical line on top of the winning
combination obtained for any of the 2 players.